requirements.txt
moviepy==1.0.3
opencv-python==4.12.0.88
pre-commit==4.3.0
PyQt5==5.15.11
python-vlc==3.0.21203
soundfile==0.13.1
sympy==1.14.0
